Refactor sample imports to new location
- Moved the Sample class to a new module.
- Updated all import statements across various files.
